+++ AnnotatoBackend/Sources/App/Models/DocumentEntity.swift
@@ -41,55 +41,3 @@ extension DocumentEntity: PersistedEntity {
     }
 }
 
-extension DocumentEntity {
-    func copyPropertiesOf(otherEntity: DocumentEntity) {
-        precondition(id == otherEntity.id)
-
-        name = otherEntity.name
-        ownerId = otherEntity.ownerId
-        baseFileUrl = otherEntity.baseFileUrl
-    }
-    
-    func loadAssociations(on db: Database) async throws {
-        try await self.$annotationEntities.load(on: db).get()
-
-        for annotationEntity in self.annotationEntities {
-            try await annotationEntity.loadAssociations(on: db)
-        }
-    }
-
-    func customUpdate(on tx: Database, usingUpdatedModel document: Document) async throws {
-        try await self.loadAssociations(on: tx)
-        try await self.pruneOldAssociations(on: tx, using: document)
-        self.copyPropertiesOf(otherEntity: DocumentEntity.fromModel(document))
-
-        for annotation in document.annotations {
-            if let annotationEntity = try await AnnotationEntity.find(annotation.id, on: tx).get() {
-                try await annotationEntity.customUpdate(on: tx, usingUpdatedModel: annotation)
-            } else {
-                try await AnnotationEntity.fromModel(annotation).create(on: tx).get()
-            }
-        }
-
-        return try await self.update(on: tx).get()
-    }
-
-
-    /// Deletes a document entity from the database. Use this function to cascade deletes.
-    /// - Parameter tx: The database instance in a transaction.
-    func customDelete(on tx: Database) async throws {
-        try await self.loadAssociations(on: tx)
-        
-        for annotationEntity in annotationEntities {
-            try await annotationEntity.customDelete(on: tx)
-        }
-
-        return try await self.delete(on: tx).get()
-    }
-
-    private func pruneOldAssociations(on tx: Database, using document: Document) async throws {
-        for annotationEntity in annotationEntities where !document.annotations.contains(where: { $0.id == annotationEntity.id }) {
-            try await annotationEntity.customDelete(on: tx)
-        }
-    }
-}
